+++ main.rs
@@ -24,6 +24,16 @@ fn plus_one(x: Option<i32>) -> Option<i32> {
 }
 
 
+//Catch-all Patterns and the _ Placeholder
+fn add_fancy_hat() {
+    println!("A fancy hat was added!");
+}
+
+fn remove_fancy_hat() {
+    println!("A fancy hat was removed!");
+}
+
+
 fn main() {
     let coin = Coin::Dime;
     let value = value_in_cents(coin);
@@ -35,4 +45,13 @@ fn main() {
     let none = plus_one(None);
     println!("six: {:?}", six);
     println!("none: {:?}", none);
+
+
+//Catch-all Patterns and the _ Placeholder
+    let dice_roll = 3;
+    match dice_roll {
+        3 => add_fancy_hat(),
+        7 => remove_fancy_hat(),
+        _ => (), 
+    }
 }